feat: add side menu visible settings
