+++ application/src/main/kotlin/viewmodels/SideMenuViewModel.kt
@@ -26,6 +26,8 @@ import models.utils.AlgorithmButton
 import models.utils.TabItem
 import themes.JetTheme
 import utils.PageType
+import utils.TextButtonAlgorithm
+import utils.TextButtonRepresentation
 import viewmodels.graphs.GraphViewModel
 import viewmodels.pages.GraphPageViewModel
 import kotlin.streams.toList
@@ -115,72 +117,5 @@ class SideMenuViewModel(graphPageViewModel: GraphPageViewModel) {
 
 	fun changeVisibility(tabName: String, isHiddenState: Boolean) {
 		sideMenuModel.changeTabVisibility(tabName, isHiddenState)
-	}
-}
-
-@Composable
-fun TextButtonAlgorithm(
-	graphViewModel: GraphViewModel?,
-	algButton: AlgorithmButton,
-	modifier: Modifier = Modifier
-) {
-	val coroutineScope = rememberCoroutineScope()
-	val expanded = mutableStateOf(false)
-	val dropContext = algButton.dropDownMenuContext
-
-	TextButton(
-		modifier = modifier,
-		onClick = {
-			if (dropContext != null) expanded.value = true
-			else algButton.isRun.value = true
-		},
-		colors = ButtonDefaults.buttonColors(
-			backgroundColor = Color.Transparent
-		)
-	) {
-		Text(algButton.label, color = JetTheme.colors.secondaryText, style = JetTheme.typography.mini)
-
-		if (dropContext != null) {
-			Box {
-				DropdownMenu(
-					modifier = Modifier
-						.background(JetTheme.colors.primaryBackground)
-						.clip(JetTheme.shapes.cornerStyle),
-					expanded = expanded.value,
-					onDismissRequest = { expanded.value = false },
-					content = {
-						Column(horizontalAlignment = Alignment.CenterHorizontally) {
-							dropContext(algButton)
-						}
-					}
-				)
-			}
-		}
-	}
-	if (graphViewModel != null && algButton.isRun.value) {
-		expanded.value = false
-		coroutineScope.launch {
-			algButton.onRun(
-				graphViewModel,
-				algButton.inputs.value.stream().map {
-					VertexID.vertexIDFromString(it, graphViewModel.vertexType)
-				}.toList()
-			)
-			algButton.isRun.value = false
-		}
-	}
-}
-
-@Composable
-fun TextButtonRepresentation(
-	graphPageViewModel: GraphPageViewModel,
-	entry: Map.Entry<String, (GraphPageViewModel) -> Unit>,
-	modifier: Modifier
-) {
-	TextButton(
-		onClick = { entry.value(graphPageViewModel) },
-		modifier = modifier
-	) {
-		Text(entry.key, color = JetTheme.colors.secondaryText, style = JetTheme.typography.mini)
 	}
 }
